Feature: use "Docker in Docker" for Kitchen tests (#2)
* feat: migrate kitchen tests to docker-in-docker

Achieving this required making a few changes to the project:

1. Use a docker driver with priveleged: true, and mounting the Kitchen
host’s docker sock
2. Create a data volume that can be mounted to both the Kitchen
container and the generated Chartmuseum container under test. Note that
this is done with docker-compose because Kitchen doesn’t seem to support
lifecycle hooks (like pre-create or post-destroy).
3. Migrate to using Rake tasks for running the Kitchen tests, which
guarantees the requisite Docker volume exists before running tests, and
ensures the volume and created Chartmuseum container is destroyed after
the tests. Note that Kitchen no longer guarantees removal of the
generated Chartmuseum Docker container since it is created as a sibling
to the Kitchen container.
4. Update the Inspec tests to extrapolate the Docker host’s IP address,
and issue the HTTP request against the bound Chartmuseum service against
that IP.

* fix: downgrade compose yml version to 3.2 (for Travis)